DRY mixin
